print_algorithm_parameters bug fix

+++ Program/AlgorithmParameters.cpp
@@ -37,6 +37,6 @@ void print_algorithm_parameters(const AlgorithmParameters & ap)
 	std::cout << "---- seed              is set to " << ap.seed << std::endl;
 	std::cout << "---- nbIter            is set to " << ap.nbIter << std::endl;
 	std::cout << "---- timeLimit         is set to " << ap.timeLimit << std::endl;
-	std::cout << "---- isRoundingInteger is set to " << ap.isRoundingInteger << std::endl;
+	std::cout << "---- isRoundingInteger is set to " << static_cast<unsigned>(ap.isRoundingInteger) << std::endl;
 	std::cout << "==================================================" << std::endl;
 }
